services/enrichment_service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.0
googlesearch-python>=1.2.0
pydantic>=2.0.0
prometheus-fastapi-instrumentator>=6.1.0
